chore: organize static assets
- Move Logo and SwapIcon components to components folder
- Extract SVG assets to public folder
- Add favicon.svg with crown logo design
- Create manifest.json
- Add robots.txt
- Remove unused vite.svg file
