+++ src/components/converter/Converter.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState } from "react";
 import type { CnbDailyRates, CnbRate } from "@/types/CnbDailyRates";
 import { CurrencyInput } from "@/components/converter/CurrencyInput";
 import { CurrencySelect } from "@/components/converter/CurrencySelect";
-import { SwapIcon } from "@/assets/SwapIcon";
+import { SwapIcon } from "@/components/SwapIcon";
 import styled from "styled-components";
 
 type Props = {
